remove redundant classes
