+++ .run/AdvancedSemiFormalStaticAnalyzerApplication.run.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="AdvancedSemiFormalStaticAnalyzerApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot" nameIsGenerated="true">
     <module name="sna" />
-    <option name="SPRING_BOOT_MAIN_CLASS" value="iu.sna.AdvancedSemiFormalStaticAnalyzerApplication" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="iu.sna.AdvancedSemiFormalStaticAnalyzerApplicationTest" />
     <method v="2">
       <option name="Maven.BeforeRunTask" enabled="true" file="$PROJECT_DIR$/pom.xml" goal="clean install" />
       <option name="Make" enabled="true" />
