+++ src/main/java/iu/sna/AdvancedSemiFormalStaticAnalyzerApplication.java
@@ -10,24 +10,12 @@ import java.io.IOException;
 
 @SpringBootApplication
 public class AdvancedSemiFormalStaticAnalyzerApplication {
-  @Value("${constants.commit_importance_coefficient}")
-  private double s;
+
 
   public static void main(String[] args) throws IOException {
     ConfigurableApplicationContext context = SpringApplication.run(
-            AdvancedSemiFormalStaticAnalyzerApplication.class, args);
+            iu.sna.AdvancedSemiFormalStaticAnalyzerApplicationTest.class, args);
 
-    AdvancedSemiFormalStaticAnalyzerApplication app =
-            context.getBean(AdvancedSemiFormalStaticAnalyzerApplication.class);
-    var graph = context.getBean(FileGraph.class);
-    graph.buildGraph();
-    var edges = graph.getEdges();
-    for (FileGraph.Edge edge : edges) {
-
-      System.out.println(edge.getFrom()
-              .getFilename() + "-- " + edge.getCompoundWeight() + " --" + edge.getTo()
-              .getFilename());
-    }
   }
 }
