--- src/main/java/iu/sna/GraphCreator/Graph.java
@@ -1,6 +1,0 @@
-package iu.sna.GraphCreator;
-
-public interface Graph {
-  FileGraph.Vertex add FileGraph.Vertex ();
-
-}
